Extract out redux from datagateway-table to common
