+++ packages/datagateway-common/src/index.tsx
@@ -7,27 +7,13 @@ export * from './app.types';
 
 export { default as Table } from './table/table.component';
 export * from './table/table.component';
-export {
-  default as DetailsPanelRow,
-} from './table/rowRenderers/detailsPanelRow.component';
-export {
-  default as DataHeader,
-} from './table/headerRenderers/dataHeader.component';
-export {
-  default as TextColumnFilter,
-} from './table/columnFilters/textColumnFilter.component';
-export {
-  default as DateColumnFilter,
-} from './table/columnFilters/dateColumnFilter.component';
-export {
-  default as ActionCellComponent,
-} from './table/cellRenderers/actionCell.component';
-export {
-  default as DataCellComponent,
-} from './table/cellRenderers/dataCell.component';
-export {
-  default as ExpandCellComponent,
-} from './table/cellRenderers/expandCell.component';
+export { default as DetailsPanelRow } from './table/rowRenderers/detailsPanelRow.component';
+export { default as DataHeader } from './table/headerRenderers/dataHeader.component';
+export { default as TextColumnFilter } from './table/columnFilters/textColumnFilter.component';
+export { default as DateColumnFilter } from './table/columnFilters/dateColumnFilter.component';
+export { default as ActionCellComponent } from './table/cellRenderers/actionCell.component';
+export { default as DataCellComponent } from './table/cellRenderers/dataCell.component';
+export { default as ExpandCellComponent } from './table/cellRenderers/expandCell.component';
 export * from './table/cellRenderers/cellContentRenderers';
 export * from './state/actions/datafiles';
 export * from './state/actions/datasets';
@@ -35,4 +21,3 @@ export * from './state/actions/investigations';
 export { default as Preloader } from './preloader/preloader.component';
 
 // ReactDOM.render(<App />, document.getElementById('root'));
-
