+++ packages/datagateway-common/src/state/actions/datasets.test.tsx
@@ -17,20 +17,11 @@ import {
   fetchDatasetDetailsRequest,
   fetchDatasetDetailsSuccess,
   fetchDatasetDetailsFailure,
-  fetchDatasetDatafilesCountRequest,
-} from '../../../datagateway-table/src/state/actions';
-import {
-  StateType,
-  EntityCache,
-} from '../../../datagateway-table/src/state/app.types';
-import { initialState } from '../../../datagateway-table/src/state/reducers/dgtable.reducer';
+} from './datasets';
+import { StateType, EntityCache } from '../app.types';
+import { initialState } from '../reducers/dgcommon.reducer';
 import axios from 'axios';
-import {
-  actions,
-  dispatch,
-  getState,
-  resetActions,
-} from '../../../datagateway-table/src/setupTests';
+import { actions, dispatch, getState, resetActions } from '../../setupTests';
 import * as log from 'loglevel';
 import { Dataset } from 'datagateway-common';
 import {
@@ -39,6 +30,8 @@ import {
   fetchDatasetSizeSuccess,
   fetchDatasetSizeFailure,
 } from './datasets';
+
+import { fetchDatasetDatafilesCountRequest } from './datafiles';
 
 jest.mock('loglevel');
 
@@ -110,7 +103,7 @@ describe('Dataset actions', () => {
   it('fetchDatasets action applies filters and sort state to request params', async () => {
     const asyncAction = fetchDatasets({ investigationId: 1 });
     const getState = (): Partial<StateType> => ({
-      dgtable: {
+      dgcommon: {
         ...initialState,
         sort: { column1: 'desc' },
         filters: { column1: '1', column2: '2' },
@@ -210,7 +203,7 @@ describe('Dataset actions', () => {
     const asyncAction = fetchDatasetSize(1);
 
     const getState = (): Partial<StateType> => ({
-      dgtable: {
+      dgcommon: {
         ...initialState,
         data: mockData,
         datasetCache: mockDatasetCache,
@@ -275,7 +268,7 @@ describe('Dataset actions', () => {
 
     const asyncAction = fetchDatasetCount(1);
     const getState = (): Partial<StateType> => ({
-      dgtable: {
+      dgcommon: {
         ...initialState,
         filters: { column1: '1', column2: '2' },
       },
@@ -401,7 +394,7 @@ describe('Dataset actions', () => {
 
     // Set up the state for calling fetchInvestigationDatasetsCountSuccess with investigation cache.
     const getState = (): Partial<StateType> => ({
-      dgtable: {
+      dgcommon: {
         ...initialState,
         data: mockData,
         investigationCache: mockInvestigationCache,
