+++ packages/datagateway-common/src/state/reducers/app.reducer.tsx
@@ -1,0 +1,16 @@
+import { combineReducers, Reducer } from 'redux';
+import dGCommonReducer from './dgcommon.reducer';
+import { connectRouter } from 'connected-react-router';
+// history package is part of react-router, which we depend on
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { History } from 'history';
+
+// import common reducer
+
+const AppReducer = (history: History): Reducer =>
+  combineReducers({
+    router: connectRouter(history),
+    dgcommon: dGCommonReducer,
+  });
+
+export default AppReducer;
