+++ packages/datagateway-table/src/state/reducers/dgtable.reducer.test.tsx
@@ -2,42 +2,13 @@ import DGTableReducer, { initialState } from './dgtable.reducer';
 import { DGTableState, EntityCache } from '../app.types';
 import {
   sortTable,
-  fetchInvestigationsRequest,
-  fetchInvestigationsSuccess,
-  fetchInvestigationsFailure,
   filterTable,
-  fetchDatasetsRequest,
-  fetchDatasetsSuccess,
-  fetchDatasetsFailure,
-  fetchDatafilesRequest,
-  fetchDatafilesSuccess,
-  fetchDatafilesFailure,
-  fetchInvestigationDatasetsCountRequest,
-  fetchInvestigationDatasetsCountSuccess,
-  fetchInvestigationDatasetsCountFailure,
-  fetchDatasetDatafilesCountRequest,
-  fetchDatasetDatafilesCountSuccess,
-  fetchDatasetDatafilesCountFailure,
   fetchInstrumentsRequest,
   fetchInstrumentsSuccess,
   fetchInstrumentsFailure,
   fetchFacilityCyclesRequest,
   fetchFacilityCyclesSuccess,
   fetchFacilityCyclesFailure,
-  downloadDatafileRequest,
-  downloadDatafileSuccess,
-  downloadDatafileFailure,
-  downloadDatasetRequest,
-  downloadDatasetSuccess,
-  downloadDatasetFailure,
-  fetchInvestigationCountRequest,
-  fetchInvestigationCountSuccess,
-  fetchInvestigationCountFailure,
-  fetchDatasetCountRequest,
-  fetchDatasetCountSuccess,
-  fetchDatasetCountFailure,
-  fetchDatafileCountRequest,
-  fetchDatafileCountFailure,
   fetchInstrumentCountRequest,
   fetchInstrumentCountSuccess,
   fetchInstrumentCountFailure,
@@ -51,18 +22,9 @@ import {
   loadUrls,
   loadBreadcrumbSettings,
   settingsLoaded,
-  fetchInvestigationDetailsRequest,
-  fetchInvestigationDetailsSuccess,
-  fetchInvestigationDetailsFailure,
-  fetchDatasetDetailsRequest,
-  fetchDatasetDetailsSuccess,
-  fetchDatasetDetailsFailure,
   fetchInstrumentDetailsRequest,
   fetchInstrumentDetailsSuccess,
   fetchInstrumentDetailsFailure,
-  fetchDatafileDetailsRequest,
-  fetchDatafileDetailsSuccess,
-  fetchDatafileDetailsFailure,
   fetchDownloadCartRequest,
   fetchDownloadCartSuccess,
   fetchDownloadCartFailure,
@@ -75,10 +37,6 @@ import {
   fetchAllIdsRequest,
   fetchAllIdsSuccess,
   fetchAllIdsFailure,
-  fetchInvestigationSizeRequest,
-  fetchInvestigationSizeSuccess,
-  fetchInvestigationSizeFailure,
-  fetchDatasetSizeSuccess,
 } from '../actions';
 import {
   Investigation,
